refactor : 보드와 연관된 책임은 ChessBoardService 로 위임
