+++ src/test/java/chess/domain/chessGame/BishopTest.java
@@ -34,15 +34,15 @@ class BishopTest {
     void generateMovablePositions() {
         Position targetPosition = new Position(Row.RANK1, Column.C);
         Color currentTurn = Color.WHITE;
-        ChessGameService chessGameService = new ChessGameService(new ChessBoardService(new MemoryBoardRepository(
+        ChessBoardService chessBoardService = new ChessBoardService(new MemoryBoardRepository(
                 Map.of(
                         targetPosition, new Piece(PieceType.BISHOP, currentTurn),
                         new Position(Row.RANK2, Column.B), new Piece(PieceType.WHITE_PAWN, currentTurn),
                         new Position(Row.RANK3, Column.E), new Piece(PieceType.BLACK_PAWN, currentTurn.opposite())
                 )
-        )), currentTurn);
+        ));
 
-        List<Position> result = chessGameService.generateMovablePositions(targetPosition);
+        List<Position> result = chessBoardService.generateMovablePositions(targetPosition, currentTurn);
 
         assertThat(result).containsExactlyInAnyOrder(
                 new Position(Row.RANK2, Column.D),
